회원정보수정 api 내부 구현 및 테스트 코드 작성

+++ app/src/main/java/com/thecommerce/app/domain/user/controller/UserApiInterface.java
@@ -76,7 +76,7 @@ public interface UserApiInterface {
                             "\t\n 3. 이메일 형식이 틀린 경우", content = @Content)
     })
     ResponseEntity<Response<UserUpdateResponseDto>> updateUser(
-            @PathVariable("id") int id,
+            @PathVariable("id") Long id,
             @Valid @RequestBody UserUpdateRequestDto userUpdateRequestDto,
             Errors errors);
 }
